requirements.txt
streamlit==1.32.0
langchain==0.1.11
langchain-openai==0.0.8
langchain-community==0.0.27
python-ffmpeg==2.0.10
Pillow==10.2.0
faiss-cpu==1.7.4